+++ app/src/main/java/com/elektro24team/auravindex/navigation/NavGraph.kt
@@ -13,13 +13,13 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
-import com.elektro24team.auravindex.utils.rememberAuditLogViewModel
-import com.elektro24team.auravindex.utils.rememberAuthViewModel
-import com.elektro24team.auravindex.utils.rememberBookCollectionViewModel
-import com.elektro24team.auravindex.utils.rememberBookViewModel
-import com.elektro24team.auravindex.utils.rememberLocalSettingViewModel
-import com.elektro24team.auravindex.utils.rememberPlanViewModel
-import com.elektro24team.auravindex.utils.rememberUserViewModel
+import com.elektro24team.auravindex.utils.functions.rememberAuditLogViewModel
+import com.elektro24team.auravindex.utils.functions.rememberAuthViewModel
+import com.elektro24team.auravindex.utils.functions.rememberBookCollectionViewModel
+import com.elektro24team.auravindex.utils.functions.rememberBookViewModel
+import com.elektro24team.auravindex.utils.functions.rememberLocalSettingViewModel
+import com.elektro24team.auravindex.utils.functions.rememberPlanViewModel
+import com.elektro24team.auravindex.utils.functions.rememberUserViewModel
 import com.elektro24team.auravindex.view.*
 import com.elektro24team.auravindex.viewmodels.AuditLogViewModel
 import com.elektro24team.auravindex.viewmodels.AuthViewModel
