+++ app/src/main/java/com/elektro24team/auravindex/utils/enums/AppAction.kt
@@ -3,7 +3,8 @@ package com.elektro24team.auravindex.utils.enums
 enum class AppAction(val appActionTitle: String) {
     SUBSCRIBE_TO_PLAN("subscribe to a plan"),
     CHECK_LISTS("check your lists"),
-    ACCESS_ADMIN_DASHBOARD("access the admin dashboard"),;
+    ACCESS_ADMIN_DASHBOARD("access the admin dashboard"),
+    ACCESS_PROFILE_PAGE("access the profile page");
 
     companion object {
         fun fromKey(appActionTitle: String): AppAction? =
