+++ app/src/main/java/com/elektro24team/auravindex/utils/functions/ExternalNavigator.kt
@@ -1,4 +1,4 @@
-package com.elektro24team.auravindex.utils
+package com.elektro24team.auravindex.utils.functions
 
 import android.content.Context
 import android.content.Intent
